feat: save persistentId when tokens are received
This commit introduces the saving of the persistent ID whenever tokens are received.
This change ensures that the persistent ID is consistently stored alongside other token-related data.
The changes were added across multiple test files and the `WelcomeScreenViewModel`.
The dependency setup for Github packages was updated in settings file to fetch all `govuk-one-login` packages from the repo.

- Updated `WelcomeScreenViewModel` to call `savePersistentId()` after updating tokens.
- Modified `WelcomeScreenTest`, `WelcomeScreenViewModelTest`, `SignedOutInfoScreenTest`, `WelcomeScreenDevMenuTest`, `WelcomeScreenDevMenuTest`, `WelcomeScreenDevMenuTest`, and `WelcomeScreenDevMenuTest` to include `savePersistentId` in constructor and mock setups.
- Updated the settings file to fetch all `govuk-one-login` packages from the Github repository.